Final report implementation
In this implementation the anchor goes to an state, after receiving the final
message, in order to send the report message. Inside the state the filtering
of the anchor is done by a switch statement. This implementation is not working very well because it is depending on the which anchor is closer.