Cooperative implementation complete
This implmenentation works with 3 anchor and 2 tags. Anchors are modified in order
to send a report message. The tags works with 2 methods. COOP and COOP_IMP. The
first depend of the other tag in order to change the role. The latter depends
on scheduling in the time. In each case the tags becomes an anchor #3. This
implementation only works with 2 tags because the limitation of the software in
only allow 4 anchors. Also, it because the limitation of hardaware. Taking into
account this(only two tags allowed), then the frequency of localization could be
ajusted in order to get more localization by second.